reorg, docs
